QuerySelect: allow optionRenderer to be passed (#947)
* QuerySelect: Allow consumers to pass the optionRendererProp
* Remove circular dependencies for QuerySelect model
* Remove circular dependencies for Lineage models
